Add support for symmetric encryption
A new flag, FLAG_SYMMETRIC is added for the EncryptText and EncryptFile
services. When specified, both will perform symmetric encryption instead of
using public keys.

To prevent programming mistakes, we error out if there is anything in the list
of recipients as it will not be used when performing symmetric encryption.

The default passphrase callback - seahorse-gpgme-source.c:passphrase_get() -
will be used to prompt users to enter the passphrase used for symmetric
encryption, like we already do for symmetric decryption.

https://bugzilla.gnome.org/show_bug.cgi?id=699161