daemon: Add a hack to find subkeys identities
The SeahorseContext does not contain PGP subkeys IDs, so is unable to
find the object corresponding to them.  This is problematic for
example for finding the identity corresponding to a signing key if
that key ID is a subkey of the primary PGP key.

For the moment, add a hack to search through all PGP keys and check
whether they have a corresponding ID when the normal hash table lookup
failed.

A better solution might be registering the subkey IDs in the context's
hash table so that the normal lookup would find the corresponding key.
However, such a change is not trivial as each module is not responsible
for registering with a specific ID but only for reporting one single ID
corresponding to the key to add.
Also, registering subkey IDs might have more deep incidence on other
code, which makes it a riskier change when not being familiar with the
code base.

https://bugzilla.gnome.org/show_bug.cgi?id=774611